+++ R/docx_exporter_functions.R
@@ -3,6 +3,23 @@ dps_markup_df_docx <- tibble::tibble(
   replace_by = c("flextable::as_sup", "flextable::as_sub")
 )
 
+align_rows_with_rtf <- function(doc) {
+  # when we have vertical pagination in the case of tables and listings,
+  # from page 2 onwards, the rows visually look a little bit shifted downwards
+  # in RTF compared to DOCX.
+  # It's a tiny difference caused by a little horizontal spacing located
+  # between the repeated table header and the first body row in each page,
+  # starting from page 2.
+  # This visual difference has to do with file "settings.xml" in the docx,
+  # and more precisely with this node:
+  # <w:compat>
+  #   <w:compatSetting w:name="compatibilityMode" w:uri="http://schemas.microsoft.com/office/word" w:val="15"/>
+  # </w:compat>
+  # we need to change the value from 15 to 11
+  doc$settings$compatibility_mode <- "11"
+  return(doc)
+}
+
 validate_tabletree <- function(tt, validate, tlgtype) {
   if (validate && tlgtype == "Table" && methods::is(tt, "VTableTree")) {
     if (!rtables::validate_table_struct(tt)) {
@@ -82,7 +99,7 @@ export_as_csv <- function(tlgtype, export_csv, pags, fontspec,
 }
 
 
-insert_fake_watermark_XML <- function(doc, watermark, orientation) {
+insert_fake_watermark_XML <- function(doc, watermark, orientation, tlgtype) {
   # This function inserts a fake watermark WordArt in the Title of
   # the flextable (that is, in the document body) by inserting an XML node.
   # This function should only be called when exporting Figures.
@@ -93,13 +110,13 @@ insert_fake_watermark_XML <- function(doc, watermark, orientation) {
   checkmate::assert_choice(orientation, choices = c("portrait", "landscape"))
 
   if (orientation == "portrait") {
-    margin_left <- -118.05
-    margin_top <- 265.35
+    margin_left <- -114
+    margin_top <- 266
   } else {
     # to shift to the right, increase margin_left
     # to shift upwards, decrease margin_top
-    margin_left <- -30.05
-    margin_top <- 175.35
+    margin_left <- -24
+    margin_top <- ifelse(tlgtype == "Listing", 165, 176)
   }
 
 
@@ -372,7 +389,7 @@ add_hanging_indent_in_title_XML <- function(doc) {
 
 add_little_gap_bottom_borders_spanning_headers <- function(
   flx,
-  border = flextable::fp_border_default(width = 0.75, color = "black")
+  border = flextable::fp_border_default(width = 0.875, color = "black")
 ) {
   nrow_header <- flextable::nrow_part(x = flx, part = "header")
   if (nrow_header > 1) {
@@ -444,7 +461,7 @@ insert_footer_text <- function(flx, tblid) {
 
 insert_title_as_header <- function(flx,
                                    title,
-                                   border = flextable::fp_border_default(width = 0.75, color = "black")) {
+                                   border = flextable::fp_border_default(width = 0.875, color = "black")) {
   # this function inserts the Title as a header but does not attempt
   # to simulate the hanging indent. Instead, it adds the string as is, and the hanging indent
   # will be handled further downstream when exporting to docx by manipulating the XML.
@@ -642,7 +659,7 @@ theme_docx_default_j <- function(
   cell_margins = c(0, 0, 0, 0),
   bold = c("header", "content_rows", "label_rows", "top_left"),
   bold_manual = NULL,
-  border = flextable::fp_border_default(width = 0.75, color = "black")
+  border = flextable::fp_border_default(width = 0.875, color = "black")
 ) {
   function(flx, ...) {
     if (!inherits(flx, "flextable")) {
@@ -787,7 +804,7 @@ tt_to_flextable_j <- function(
     tt,
     tblid = NULL,
     theme = theme_docx_default_j(font = "Times New Roman", font_size = 9L, bold = NULL),
-    border = flextable::fp_border_default(width = 0.75, color = "black"),
+    border = flextable::fp_border_default(width = 0.875, color = "black"),
     titles_as_header = TRUE,
     bold_titles = TRUE,
     integrate_footers = TRUE,
@@ -1512,7 +1529,7 @@ export_as_docx_j <- function(
   pagenum = ifelse(tlgtype == "Listing", TRUE, FALSE),
   round_type = formatters::obj_round_type(tt),
   alignments = list(),
-  border = flextable::fp_border_default(width = 0.75, color = "black"),
+  border = flextable::fp_border_default(width = 0.875, color = "black"),
   border_mat = make_header_bordmat(obj = tt),
   watermark = NULL,
   export_csv = FALSE,
@@ -1842,8 +1859,10 @@ export_as_docx_j <- function(
     add_vertical_pagination_XML(doc)
     remove_security_popup_page_numbers_XML(doc, tlgtype, pagenum)
     if (!is.null(watermark)) {
-      insert_fake_watermark_XML(doc, watermark, orientation)
+      insert_fake_watermark_XML(doc, watermark, orientation, tlgtype)
     }
+
+    doc <- align_rows_with_rtf(doc)
 
     print(doc, target = paste0(output_dir, "/", tolower(tblid), ".docx"))
     invisible(TRUE)
@@ -1880,7 +1899,7 @@ export_graph_as_docx <- function(g = NULL,
                                  plotwidth = 8,
                                  plotheight = 5.51,
                                  units = c("in", "cm", "mm", "px")[1],
-                                 border = flextable::fp_border_default(width = 0.75, color = "black"),
+                                 border = flextable::fp_border_default(width = 0.875, color = "black"),
                                  watermark = NULL) {
 
   # Check arguments ----
@@ -2051,7 +2070,7 @@ export_graph_as_docx <- function(g = NULL,
   doc <- flextable::body_add_flextable(doc, flx, align = "center")
   add_hanging_indent_in_title_XML(doc)
   if (!is.null(watermark)) {
-    insert_fake_watermark_XML(doc, watermark, orientation)
+    insert_fake_watermark_XML(doc, watermark, orientation, "Figure")
   }
   print(doc, target = paste0(output_dir, "/", tolower(tblid), ".docx"))
 }
@@ -2118,7 +2137,7 @@ export_graph_as_docx <- function(g = NULL,
 #' Each named list contains `row`, `col`, and `value`.\cr
 #' (optional) Default = `list()`.
 #' @param border (`fp_border`)\cr border to use.\cr
-#' Default = \code{flextable::fp_border_default(width = 0.75, color = "black")}.
+#' Default = \code{flextable::fp_border_default(width = 0.875, color = "black")}.
 #' @param border_mat (`matrix`)\cr a `m x k` matrix where m is the number of columns of
 #' the input Table/Listing and k is the number of lines the header takes up.\cr
 #' See [tidytlg::add_bottom_borders] for what the matrix should contain.
@@ -2210,7 +2229,7 @@ export_graph_as_docx <- function(g = NULL,
 #'   pagenum = FALSE,
 #'   round_type = "iec",
 #'   alignments = list(),
-#'   border = flextable::fp_border_default(width = 0.75, color = "black"),
+#'   border = flextable::fp_border_default(width = 0.875, color = "black"),
 #'   border_mat = NULL,
 #'   watermark = NULL,
 #'   plotnames = NULL,
@@ -2265,7 +2284,7 @@ export_TLG_as_docx <- function(
                       formatters::obj_round_type(obj),
                       "iec"),
   alignments = list(),
-  border = flextable::fp_border_default(width = 0.75, color = "black"),
+  border = flextable::fp_border_default(width = 0.875, color = "black"),
   border_mat = NULL,
   export_csv = FALSE,
   output_csv_directory = NULL,